my second edition

+++ assignment1/src/eg/edu/alexu/csd/datastructure/calculator/cs59/MyCalculator.java
@@ -9,7 +9,7 @@ public class MyCalculator implements ICalculator {
 
 	
 	public float divide(int x, int y) {
-		return (float)x/(float)y;
+		return (float)x/y;
 	}
 
 }
